release version 5.1.1

+++ app/src/main/java/com/voidapp/magizhiniorganics/magizhiniorganics/ui/checkout/InvoiceActivity.kt
@@ -656,7 +656,7 @@ class InvoiceActivity :
             binding.tvTotalAmt.setTextAnimation(
                 "${
                     totalPrice + viewModel.gstAmount + binding.tvDeliveryChargeAmt.text.toString()
-                        .toFloat()
+                        .toDouble()
                 }"
             )
